config overhaul / switch to yaml
