+++ tud_rl/run/train_discrete.py
@@ -9,36 +9,28 @@ import gym_minatar
 import gym_pygame
 import numpy as np
 import torch
+from tud_rl.agents.base import AbstractAgent
+from tud_rl.common import configparser
 from tud_rl.envs.MountainCar import MountainCar
 from tud_rl.envs.FossenEnv import *
-from tud_rl.wrappers.MinAtar_wrapper import MinAtar_wrapper
-from tud_rl.agents.discrete.BootDQN import BootDQNAgent
-from tud_rl.agents.discrete.DDQN import DDQNAgent
-from tud_rl.agents.discrete.DQN import DQNAgent
-from tud_rl.agents.discrete.ACCDDQN import ACCDDQNAgent
-from tud_rl.agents.discrete.EnsembleDQN import EnsembleDQNAgent
-from tud_rl.agents.discrete.KEBootDQN import KEBootDQNAgent
-from tud_rl.agents.discrete.KEEnsembleDQN import KEEnsembleDQNAgent
-from tud_rl.agents.discrete.MaxMinDQN import MaxMinDQNAgent
-from tud_rl.agents.discrete.SCDQN import SCDQNAgent
-from tud_rl.agents.discrete.RecDQN import RecDQNAgent
-from tud_rl.agents.discrete.ComboDQN import ComboDQNAgent
+from tud_rl.wrappers import get_wrapper
+from tud_rl.agents.discrete import *
 from tud_rl.common.logging_plot import plot_from_progress
 from tud_rl.configs.discrete_actions import __path__
 
 
-def evaluate_policy(test_env, agent, c):
+def evaluate_policy(test_env: gym.Env, agent: BaseAgent, c: Configfile):
 
     # go greedy
     agent.mode = "test"
     
     rets = []
     
-    for _ in range(c["eval_episodes"]):
+    for _ in range(c.eval_episodes):
 
         # get initial state
         s = test_env.reset()
-        if c["input_norm"]:
+        if c.input_norm:
             s = agent.inp_normalizer.normalize(s, mode=agent.mode)
 
         cur_ret = 0
@@ -56,7 +48,7 @@ def evaluate_policy(test_env, agent, c):
             s2, r, d, _ = test_env.step(a)
 
             # potentially normalize s2
-            if c["input_norm"]:
+            if c.input_norm:
                 s2 = agent.inp_normalizer.normalize(s2, mode=agent.mode)
 
             # s becomes s2
@@ -64,7 +56,7 @@ def evaluate_policy(test_env, agent, c):
             cur_ret += r
 
             # break option
-            if eval_epi_steps == c["env"]["max_episode_steps"]:
+            if eval_epi_steps == c.Env.max_episode_steps:
                 break
         
         # append return
@@ -76,53 +68,53 @@ def evaluate_policy(test_env, agent, c):
     return rets
 
 
-def train(c, agent_name):
+def train(c: Configfile, agent_name: str):
     """Main training loop."""
 
     # measure computation time
     start_time = time.time()
     
     # init envs
-    env = gym.make(c["env"]["name"], **c["env"]["env_kwargs"])
-    test_env = gym.make(c["env"]["name"], **c["env"]["env_kwargs"])
+    env = gym.make(c.Env.name, **c.Env.env_kwargs)
+    test_env = gym.make(c.Env.name, **c.Env.env_kwargs)
 
     # wrappers
-    for wrapper in c["env"]["wrappers"]:
-        wrapper_kwargs = c["env"]["wrapper_kwargs"][wrapper]
-        env = eval(wrapper)(env, **wrapper_kwargs)
-        test_env = eval(wrapper)(test_env, **wrapper_kwargs)
+    for wrapper in c.Env.wrappers:
+        wrapper_kwargs = c.Env.wrapper_kwargs[wrapper]
+        env: gym.Env = get_wrapper(wrapper, env, **wrapper_kwargs)
+        test_env: gym.Env = get_wrapper(wrapper,test_env, **wrapper_kwargs)
 
     # get state_shape
-    if c["env"]["state_type"] == "image":
-        if "MinAtar" in c["env"]["name"]:
+    if c.Env.state_type == "image":
+        if "MinAtar" in c.Env.name:
             # careful, MinAtar constructs state as (height, width, in_channels), which is NOT aligned with PyTorch
-            c["state_shape"] = (env.observation_space.shape[2], *env.observation_space.shape[0:2])
+            c.state_shape = (env.observation_space.shape[2], *env.observation_space.shape[0:2])
         else:
-            c["state_shape"] = env.observation_space.shape[0]
+            c.state_shape = env.observation_space.shape[0]
     
-    elif c["env"]["state_type"] == "feature":
-        c["state_shape"] = env.observation_space.shape[0]
+    elif c.Env.state_type == "feature":
+        c.state_shape = env.observation_space.shape[0]
 
     # mode and num actions
-    c["mode"] = "train"
-    c["num_actions"] = env.action_space.n
+    c.mode = "train"
+    c.num_actions = env.action_space.n
 
     # seeding
-    env.seed(c["seed"])
-    test_env.seed(c["seed"])
-    torch.manual_seed(c["seed"])
-    np.random.seed(c["seed"])
-    random.seed(c["seed"])
+    env.seed(c.seed)
+    test_env.seed(c.seed)
+    torch.manual_seed(c.seed)
+    np.random.seed(c.seed)
+    random.seed(c.seed)
 
     # init agent
     if agent_name[-1].islower():
-        agent = eval(agent_name[:-2] + "Agent")(c, agent_name)
+        agent: AbstractAgent = eval(agent_name[:-2] + "Agent")(c, agent_name)
     else:
-        agent = eval(agent_name + "Agent")(c, agent_name)
+        agent: AbstractAgent = eval(agent_name + "Agent")(c, agent_name)
 
     # get initial state and normalize it
     s = env.reset()
-    if c["input_norm"]:
+    if c.input_norm:
         s = agent.inp_normalizer.normalize(s, mode=agent.mode)
 
     # init epi step counter and epi return
@@ -130,12 +122,12 @@ def train(c, agent_name):
     epi_ret = 0
     
     # main loop    
-    for total_steps in range(c["timesteps"]):
+    for total_steps in range(c.timesteps):
 
         epi_steps += 1
         
         # select action
-        if total_steps < c["act_start_step"]:
+        if total_steps < c.act_start_step:
             a = np.random.randint(low=0, high=agent.num_actions, size=1, dtype=int).item()
         else:
             a = agent.select_action(s)
@@ -144,10 +136,10 @@ def train(c, agent_name):
         s2, r, d, _ = env.step(a)
         
         # Ignore "done" if it comes from hitting the time horizon of the environment
-        d = False if epi_steps == c["env"]["max_episode_steps"] else d
+        d = False if epi_steps == c.Env.max_episode_steps else d
 
         # potentially normalize s2
-        if c["input_norm"]:
+        if c.input_norm:
             s2 = agent.inp_normalizer.normalize(s2, mode=agent.mode)
 
         # add epi ret
@@ -157,15 +149,15 @@ def train(c, agent_name):
         agent.memorize(s, a, r, s2, d)
 
         # train
-        if (total_steps >= c["upd_start_step"]) and (total_steps % c["upd_every"] == 0):
-            for _ in range(c["upd_every"]):
+        if (total_steps >= c.upd_start_step) and (total_steps % c.upd_every == 0):
+            for _ in range(c.upd_every):
                 agent.train()
 
         # s becomes s2
         s = s2
 
         # end of episode handling
-        if d or (epi_steps == c["env"]["max_episode_steps"]):
+        if d or (epi_steps == c.Env.max_episode_steps):
  
             # reset active head for BootDQN
             if "Boot" in agent_name:
@@ -173,7 +165,7 @@ def train(c, agent_name):
 
             # reset to initial state and normalize it
             s = env.reset()
-            if c["input_norm"]:
+            if c.input_norm:
                 s = agent.inp_normalizer.normalize(s, mode=agent.mode)
             
             # log episode return
@@ -184,9 +176,9 @@ def train(c, agent_name):
             epi_ret = 0
 
         # end of epoch handling
-        if (total_steps + 1) % c["epoch_length"] == 0 and (total_steps + 1) > c["upd_start_step"]:
+        if (total_steps + 1) % c.epoch_length == 0 and (total_steps + 1) > c.upd_start_step:
 
-            epoch = (total_steps + 1) // c["epoch_length"]
+            epoch = (total_steps + 1) // c.epoch_length
 
             # evaluate agent with deterministic policy
             eval_ret = evaluate_policy(test_env=test_env, agent=agent, c=c)
@@ -204,14 +196,18 @@ def train(c, agent_name):
             agent.logger.dump_tabular()
 
             # create evaluation plot based on current 'progress.txt'
-            plot_from_progress(dir=agent.logger.output_dir, alg=agent.name, env_str=c["env"]["name"], info=c["env"]["info"])
+            plot_from_progress(
+                dir=agent.logger.output_dir, 
+                alg=agent.name, 
+                env_str=c.Env.name, 
+                info=c.Env.info)
 
             # save weights
             if not any([word in agent.name for word in ["ACCDDQN", "Ensemble", "MaxMin"]]):
                 torch.save(agent.DQN.state_dict(), f"{agent.logger.output_dir}/{agent.name}_weights.pth")
     
             # save input normalizer values 
-            if c["input_norm"]:
+            if c.input_norm:
                 with open(f"{agent.logger.output_dir}/{agent.name}_inp_norm_values.pickle", "wb") as f:
                     pickle.dump(agent.inp_normalizer.get_for_save(), f)
 
@@ -220,34 +216,25 @@ if __name__ == "__main__":
 
     # get config and name of agent
     parser = argparse.ArgumentParser()
-    parser.add_argument("--config_file", type=str, default="pathfollower.json")
+    parser.add_argument("--config_file", type=str, default="pathfollower.yaml")
     parser.add_argument("--lr", type=float, default=None)
     parser.add_argument("--seed", type=int, default=None)
-    parser.add_argument("--agent_name", type=str, default="DQN")
+    parser.add_argument("--agent_name", type=str, default="SCDQN_b")
     args = parser.parse_args()
 
-    # read config file
-    with open(__path__._path[0] + "/" + args.config_file) as f:
-        c = json.load(f)
+    config_path = __path__._path[0] + "/" + args.config_file
+    
+    # Parse the config file into a python class to avoid magic strings
+    config = configparser.Configfile(config_path)
 
     # potentially overwrite lr and seed
     if args.lr is not None:
-        c["lr"] = args.lr
+        config.lr = args.lr
     if args.seed is not None:
-        c["seed"] = args.seed
-
-    # convert certain keys in integers
-    for key in ["seed", "timesteps", "epoch_length", "eval_episodes", "eps_decay_steps", "tgt_update_freq",\
-        "buffer_length", "act_start_step", "upd_start_step", "upd_every", "batch_size"]:
-        c[key] = int(c[key])
+        config.seed = args.seed
 
     # handle maximum episode steps
-    if c["env"]["max_episode_steps"] == -1:
-        c["env"]["max_episode_steps"] = np.inf
-    else:
-        c["env"]["max_episode_steps"] = int(c["env"]["max_episode_steps"])
+    if config.Env.max_episode_steps == -1:
+        config.Env.max_episode_steps = np.inf
 
-    # set number of torch threads
-    torch.set_num_threads(torch.get_num_threads())
-
-    train(c, args.agent_name)
+    train(config, args.agent_name)
